Add some minitest helpers for working with temp files & dirs, as well as fixture files.
It's nice to be able to just ask for a temp file or directory, and
have it automatically removed at the end of the test, so add #tempdir
and #tempfile helpers that automatically register the director and
file path to be removed as part of teardown.

This also adds a couple of helpers around copying fixture files into
place, and reading their contents, since it's nice to not have to
specify the full paths everywhere.