Start Hummingbird::Database
Hummingbird::Database will be responsible for all direct interaction
with the database, including listing the migrations that have already
been run, as well as running new migrations.

So far, the only implemented functionality is reporting whether the
database has been initialized for use by Hummingbird or not, and
listing the migrations that have already been run against the
database.