Do not modify Plan#planned_files from #to_be_run_migration_file_names
Use a duplicate of #planned_files from within #to_be_run_migration_file_names
rather than the original, since we're modifying it to get the list of
files that have yet to be run.  This change allows people to always be
able to get the entire list of planned files, even after we've
computed the ones that have already been run.