Allow overriding the "general" and "user" file names
This change also removes the assumption that the user config file will
be in FileUtils.pwd, and instead assumes that it will be in the
specified configuration directory.

With this change it is also possible to have the config files in
subdirectories of the configuration directory, or outside of it
entirely (by either using '..' to walk up the tree, or by using an
absolute path).